requirements.txt
scrapy
requests
pandas
curl_cffi
mysql-connector-python
python-dotenv